requirements.txt
Django==3.2.22
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
requests==2.31.0
django-grappelli==3.0.2
drf-yasg==1.20.0
gunicorn==20.1.0
supervisor==4.2.4
psycopg2==2.9.3
django-rest-knox==4.2.0
